requirements.txt
pillow==10.4.0
bcrypt==3.2.0
tk==0.1.0
